requirements.txt
Django==1.11.29
django-tastypie==0.14.1
python-dateutil==2.9.0.post0
python-mimeparse==1.6.0
pytz==2024.2
six==1.16.0
